% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cprod}
\alias{cprod}
\title{Colsums of crossprod}
\usage{
cprod(X, Y, summit, transposeX, transposeY)
}
\arguments{
\item{X}{A matrix with dimensions k*n.}

\item{Y}{A matrix with dimenions n*m.}

\item{summit}{Logical. If \code{TRUE} return colsums of the resulting matrix.}

\item{transposeX}{Logical. If \code{TRUE} transpose X before multiplication.}

\item{transposeY}{Logical. If \code{TRUE} transpose Y before multiplication.}
}
\value{
If \code{summit=0} a matrix with dimensions k * m else a vector of length m.
}
\description{
Fast computation of X %*% Y and colSums(X%*%Y)
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}

