% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_assessed_20251001}
\alias{cran_assessed_20251001}
\title{CRAN riskmetric Assessments on 2025-10-01}
\format{
A \code{data.frame}

\describe{
\item{package}{The name of a R package hosted on CRAN}
\item{version}{The package's version number}
\item{riskmetric_run_date}{The date riskmetric was run}
\item{riskmetric_version}{The version of riskmetric used to derive scores}
}
All other variables correspond to the riskmetric assessments that exist at
the time of execution. Please refer to riskmetric's documentation for more
info.
}
\usage{
cran_assessed_20251001
}
\value{
A data.frame
}
\description{
A data.frame containing all of CRAN was assessed using riskmetric v0.2.5 on
2025-10-01 & R v4.5.1.
}
\keyword{datasets}
