% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{print.Score}
\alias{print.Score}
\title{Print Score object}
\usage{
\method{print}{Score}(x, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{...}{passed to print}
}
\description{
Print method for risk prediction scores
}

