% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_risk.R
\name{pull_risk}
\alias{pull_risk}
\title{Pull Risk Data}
\usage{
pull_risk(source = "cran", type = c("assessed", "scored")[1], date = "latest")
}
\arguments{
\item{source}{character}

\item{type}{character, either 'assessed' or 'scored'}

\item{date}{character, of format 'YYYY-MM-DD',  'YYYYMMDD', or 'latest'}
}
\value{
A data.frame
}
\description{
A function to pull risk assessment & / or risk scores for all CRAN packages
}
\examples{
pull_risk(type = "scored")

}
\keyword{internal}
