% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskScore-package.R
\docType{package}
\name{riskScore}
\alias{riskScore}
\alias{riskScore-package}
\title{Functions for scoring risk markers and risk prediction models}
\description{
riskScore allows you to assess many markers/models simultaneously
}

