% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{Score}
\alias{Score}
\title{Score risk predictions}
\usage{
Score(object, ...)
}
\arguments{
\item{object}{see Score.list}

\item{...}{see Score.list}
}
\value{
see Score.list
}
\description{
Generic function to Score risk predictions
}
\author{
Thomas Alexander Gerds
}

