% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Score.R
\name{print.Score}
\alias{print.Score}
\title{Print Scores and tests risk predictions}
\usage{
\method{print}{Score}(x, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{...}{passed to print}
}
\description{
Print of Scored risk predictions
}

