% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{print.score}
\alias{print.score}
\title{Print metric specific results of risk prediction assessment}
\usage{
\method{print}{score}(x, B = 0, ...)
}
\arguments{
\item{x}{Element of the result of \code{Score.list}}

\item{B}{Number of splits if any.}

\item{...}{passed to print}
}
\description{
Print metric specific element of risk prediction assessment
}

