% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-initial-placeholder.R
\docType{data}
\name{v0.2.1_cran_20230621}
\alias{v0.2.1_cran_20230621}
\title{Initial Placeholder Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19715 rows and 24 columns.
}
\usage{
v0.2.1_cran_20230621
}
\value{
A data.frame
}
\description{
The initial data.frame is generated as a placeholder until a more streamlined
process is implemented. It contains the initial run where all of CRAN was
scored using riskmetric v0.2.1 on 2023-06-21. It contains 19,715 observations
(one row per package) and 24 variables, 18 of which are riskmetric
assessments.
}
\keyword{datasets}
